toolkit/content/tests/chrome/frame_subframe_origin_subframe1.xul

<?xml version="1.0"?>
<!-- ***** BEGIN LICENSE BLOCK *****
   - Version: MPL 1.1/GPL 2.0/LGPL 2.1
   -
   - The contents of this file are subject to the Mozilla Public License Version
   - 1.1 (the "License"); you may not use this file except in compliance with
   - the License. You may obtain a copy of the License at
   - http://www.mozilla.org/MPL/
   -
   - Software distributed under the License is distributed on an "AS IS" basis,
   - WITHOUT WARRANTY OF ANY KIND, either express or implied. See the License
   - for the specific language governing rights and limitations under the
   - License.
   -
   - The Original Code is Findbar Test code
   -
   - The Initial Developer of the Original Code is
   - Mozilla Foundation.
   - Portions created by the Initial Developer are Copyright (C) 2010
   - the Initial Developer. All Rights Reserved.
   -
   - Contributor(s):
   -
   - Alternatively, the contents of this file may be used under the terms of
   - either the GNU General Public License Version 2 or later (the "GPL"), or
   - the GNU Lesser General Public License Version 2.1 or later (the "LGPL"),
   - in which case the provisions of the GPL or the LGPL are applicable instead
   - of those above. If you wish to allow use of your version of this file only
   - under the terms of either the GPL or the LGPL, and not to allow others to
   - use your version of this file under the terms of the MPL, indicate your
   - decision by deleting the provisions above and replace them with the notice
   - and other provisions required by the GPL or the LGPL. If you do not delete
   - the provisions above, a recipient may use your version of this file under
   - the terms of any one of the MPL, the GPL or the LGPL.
   -
   - ***** END LICENSE BLOCK ***** -->
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>

<page id="frame1"
      style="background-color:green;" 
      xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
<script type="application/javascript"
        src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"/>

<spacer height="10px"/>
<iframe
  style="margin:10px; min-height:170px; max-width:200px; max-height:200px; border:solid 1px white;"
  src="frame_subframe_origin_subframe2.xul"></iframe>
<spacer height="3px"/>
<caption id="cap1" style="min-width:200px; max-width:200px; background-color:white;" label=""/>
<script class="testbody" type="application/javascript">

// Fire a mouse move event aimed at this window, and check to be
// sure the client coords translate from widget to the dom correctly.

function runTests()
{
  synthesizeMouse(document.getElementById("frame1"), 3, 4, { type: "mousemove" });
}

function mouseMove(e) {
  e.stopPropagation();
  var element = e.target;
  var el = document.getElementById("cap1");
  el.label = "client: (" + e.clientX + "," + e.clientY + ")";
  parent.opener.wrappedJSObject.SimpleTest.is(e.clientX, 3, "mouse event clientX on sub frame 1");
  parent.opener.wrappedJSObject.SimpleTest.is(e.clientY, 4, "mouse event clientY on sub frame 1");
  // fire the next test on the sub frame
  frames[0].runTests();
}

window.addEventListener("mousemove", mouseMove, false);

</script>
</page>
